+++ packages/riverpod/test/src/providers/stream_notifier_test.dart
@@ -44,7 +44,7 @@ void main() {
           final listener = Listener<AsyncValue<int>>();
 
           container.listen(provider, fireImmediately: true, listener.call);
-          await container.read(provider.future).catchError((e) => 0);
+          await container.read(provider.future).sync.catchError((e) => 0);
 
           verifyOnly(
             listener,
@@ -56,7 +56,7 @@ void main() {
           fake.elapse(const Duration(seconds: 1));
           fake.flushMicrotasks();
 
-          await container.read(provider.future).catchError((e) => 0);
+          await container.read(provider.future).sync.catchError((e) => 0);
 
           verifyOnly(
             listener,
@@ -175,7 +175,7 @@ void main() {
 
         container.listen(provider, (previous, next) {});
 
-        await expectLater(container.read(provider.future), completion(0));
+        await expectLater(container.read(provider.future), completionOr(0));
         expect(container.read(provider), const AsyncData(0));
 
         expect(
@@ -183,7 +183,7 @@ void main() {
           const AsyncLoading<int>().copyWithPrevious(const AsyncData(0)),
         );
 
-        await expectLater(container.read(provider.future), completion(1));
+        await expectLater(container.read(provider.future), completionOr(1));
         expect(container.read(provider), const AsyncData(1));
 
         container.invalidate(provider);
@@ -192,7 +192,7 @@ void main() {
           container.read(provider),
           const AsyncLoading<int>().copyWithPrevious(const AsyncData(1)),
         );
-        await expectLater(container.read(provider.future), completion(2));
+        await expectLater(container.read(provider.future), completionOr(2));
         expect(container.read(provider), const AsyncData(2));
       });
 
@@ -204,7 +204,7 @@ void main() {
 
         final sub = container.listen(provider.notifier, (previous, next) {});
 
-        await expectLater(container.read(provider.future), completion(0));
+        await expectLater(container.read(provider.future), completionOr(0));
         expect(container.read(provider), const AsyncData(0));
 
         sub.read().state = const AsyncLoading<int>();
@@ -227,7 +227,7 @@ void main() {
 
         container.listen(provider, (previous, next) {});
 
-        await expectLater(container.read(provider.future), completion(0));
+        await expectLater(container.read(provider.future), completionOr(0));
         expect(container.read(provider), const AsyncData(0));
 
         container.read(dep.notifier).state++;
@@ -237,7 +237,7 @@ void main() {
               .copyWithPrevious(const AsyncData(0), isRefresh: false),
         );
 
-        await expectLater(container.read(provider.future), completion(1));
+        await expectLater(container.read(provider.future), completionOr(1));
         expect(container.read(provider), const AsyncData(1));
       });
 
@@ -252,7 +252,7 @@ void main() {
 
         container.listen(provider, (previous, next) {});
 
-        await expectLater(container.read(provider.future), completion(0));
+        await expectLater(container.read(provider.future), completionOr(0));
         expect(container.read(provider), const AsyncData(0));
 
         container.read(dep.notifier).state++;
@@ -262,7 +262,7 @@ void main() {
               .copyWithPrevious(const AsyncData(0), isRefresh: false),
         );
 
-        await expectLater(container.read(provider.future), completion(1));
+        await expectLater(container.read(provider.future), completionOr(1));
         expect(container.read(provider), const AsyncData(1));
       });
     });
@@ -432,7 +432,7 @@ void main() {
 
       expect(
         container.read(provider.future),
-        completion(21),
+        completionOr(21),
         reason: 'The provider rebuilt while the future was still pending, '
             'so .future should resolve with the next value',
       );
@@ -444,7 +444,7 @@ void main() {
 
       verifyZeroInteractions(listener);
 
-      expect(container.read(provider.future), completion(21));
+      expect(container.read(provider.future), completionOr(21));
       expect(container.read(provider), const AsyncLoading<int>());
 
       completers[1]!.complete(21);
@@ -602,7 +602,7 @@ void main() {
 
         completer.complete(42);
 
-        await expectLater(future, completion(42));
+        await expectLater(future, completionOr(42));
       });
 
       test(
@@ -643,7 +643,7 @@ void main() {
 
           completer.complete(42);
 
-          await expectLater(future, completion(42));
+          await expectLater(future, completionOr(42));
         },
       );
 
@@ -679,24 +679,24 @@ void main() {
         final provider = factory.simpleTestProvider<int>(
           (ref) => Stream.fromFuture(completer.future),
         );
-        final listener = Listener<Future<int>>();
+        final listener = Listener<FutureOr<int>>();
 
         final sub = container.listen(provider.notifier, (previous, next) {});
         container.listen(provider.future, listener.call);
 
-        expect(sub.read().future, completion(21));
+        expect(sub.read().future, completionOr(21));
 
         sub.read().state = const AsyncData(21);
 
         completer.complete(42);
 
-        expect(sub.read().future, completion(42));
+        expect(sub.read().future, completionOr(42));
         final capture =
             verifyOnly(listener, listener(captureAny, captureAny)).captured;
 
         expect(capture.length, 2);
-        expect(capture.first, completion(21));
-        expect(capture.last, completion(42));
+        expect(capture.first, completionOr(21));
+        expect(capture.last, completionOr(42));
       });
 
       test('resolves with the new state when notifier.state is changed',
@@ -705,7 +705,7 @@ void main() {
         final provider = factory.simpleTestProvider<int>(
           (ref) => Stream.value(0),
         );
-        final listener = Listener<Future<int>>();
+        final listener = Listener<FutureOr<int>>();
 
         final sub = container.listen(provider.notifier, (previous, next) {});
         container.listen(
@@ -714,15 +714,15 @@ void main() {
           fireImmediately: true,
         );
 
-        await expectLater(sub.read().future, completion(0));
+        await expectLater(sub.read().future, completionOr(0));
         verifyOnly(
           listener,
-          listener(argThat(equals(null)), argThat(completion(0))),
+          listener(argThat(equals(null)), argThat(completionOr(0))),
         );
 
         sub.read().state = const AsyncData(1);
 
-        await expectLater(sub.read().future, completion(1));
+        await expectLater(sub.read().future, completionOr(1));
       });
 
       test('returns a Future identical to that of .future', () {
@@ -831,15 +831,15 @@ void main() {
 
         await expectLater(
           sub.read().update((prev) => prev + 1),
-          completion(1),
+          completionOr(1),
         );
         await expectLater(
           sub.read().future,
-          completion(1),
+          completionOr(1),
         );
         await expectLater(
           sub.read().update((prev) => prev + 1),
-          completion(2),
+          completionOr(2),
         );
       });
 
@@ -871,7 +871,7 @@ void main() {
               return 21;
             },
           ),
-          completion(21),
+          completionOr(21),
         );
         expect(callCount, 0);
         expect(actualErr, 42);
@@ -895,7 +895,7 @@ void main() {
 
         await expectLater(
           sub.read().update((prev) => prev + 1),
-          completion(2),
+          completionOr(2),
         );
         expect(container.read(provider), const AsyncData(2));
       });
@@ -1046,8 +1046,8 @@ void main() {
       canBeAssignedToProviderListenable<AsyncValue<int>>(provider);
       canBeAssignedToRefreshable<AsyncValue<int>>(provider);
 
-      canBeAssignedToProviderListenable<Future<int>>(provider.future);
-      canBeAssignedToRefreshable<Future<int>>(provider.future);
+      canBeAssignedToProviderListenable<FutureOr<int>>(provider.future);
+      canBeAssignedToRefreshable<FutureOr<int>>(provider.future);
 
       canBeAssignedToProviderListenable<StreamNotifier<int>>(provider.notifier);
       canBeAssignedToRefreshable<StreamNotifier<int>>(provider.notifier);
@@ -1065,8 +1065,8 @@ void main() {
       canBeAssignedToProviderListenable<AsyncValue<int>>(autoDispose);
       canBeAssignedToRefreshable<AsyncValue<int>>(autoDispose);
 
-      canBeAssignedToProviderListenable<Future<int>>(autoDispose.future);
-      canBeAssignedToRefreshable<Future<int>>(autoDispose.future);
+      canBeAssignedToProviderListenable<FutureOr<int>>(autoDispose.future);
+      canBeAssignedToRefreshable<FutureOr<int>>(autoDispose.future);
 
       canBeAssignedToProviderListenable<StreamNotifier<int>>(
         autoDispose.notifier,
@@ -1088,8 +1088,8 @@ void main() {
       canBeAssignedToProviderListenable<AsyncValue<String>>(family(0));
       canBeAssignedToRefreshable<AsyncValue<String>>(family(0));
 
-      canBeAssignedToProviderListenable<Future<String>>(family(0).future);
-      canBeAssignedToRefreshable<Future<String>>(family(0).future);
+      canBeAssignedToProviderListenable<FutureOr<String>>(family(0).future);
+      canBeAssignedToRefreshable<FutureOr<String>>(family(0).future);
 
       canBeAssignedToProviderListenable<FamilyStreamNotifier<String, int>>(
         family(0).notifier,
@@ -1120,10 +1120,10 @@ void main() {
         autoDisposeFamily(0),
       );
 
-      canBeAssignedToProviderListenable<Future<String>>(
+      canBeAssignedToProviderListenable<FutureOr<String>>(
         autoDisposeFamily(0).future,
       );
-      canBeAssignedToRefreshable<Future<String>>(
+      canBeAssignedToRefreshable<FutureOr<String>>(
         autoDisposeFamily(0).future,
       );