+++ website/docs/case_studies/cancel/detail_screen/raw.dart
@@ -49,7 +49,7 @@ class DetailPageView extends ConsumerWidget {
         title: const Text('Detail page'),
       ),
       body: RefreshIndicator(
-        onRefresh: () => ref.refresh(activityProvider.future),
+        onRefresh: () async => ref.refresh(activityProvider.future),
         child: ListView(
           children: [
             switch (activity) {